+++ changelog.md
@@ -35,7 +35,8 @@
 - Fixed alternative cybergrind loot tables for enemies being messed up, resulting in nothing being dropped at all
 - Made getting locked out of the `destiny-chapel` Limbo Challenge Structure harder
 - Hopefully fixed Structures and Levels being breakable when they shouldn't be
-- Fixed typo in Triple Kill 
+- Fixed typo in Triple Kill
+- Fixed Main Menu Crashing the game when using Vivecraft
 ## Settings & Config
 ## Commands
 ## Tweaks